requirements.txt
bleach==1.5.0
boost==0.1
certifi==2017.7.27.1
chardet==3.0.4
Cython==0.27.1
Flask>=0.11.0
h5py==2.7.1
html5lib==0.9999999
idna==2.6
Markdown==2.6.9
Mastodon.py==1.1.1
numpy>=1.13.1
olefile==0.44
pandas==0.20.3
Pillow==4.2.1
protobuf==3.4.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.4
scikit-learn==0.19.0
scipy==0.19.1
six==1.11.0
SQLAlchemy==1.1.14
tensorflow-tensorboard==0.1.7
urllib3==1.22
Werkzeug==0.12.2